+++ changelog.md
@@ -19,3 +19,4 @@
 - Added `jssets` module, Set for the JavaScript target
   https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Set
 - Added `jsheaders` module for [`Headers`](https://developer.mozilla.org/en-US/docs/Web/API/Headers) for the JavaScript target.
+- Added module `pointers` containing `toUncheckedArray`, UFCS/MCS `cast`